LTR iterator relocated
