+++ src/main/java/com/mmm/his/cer/utility/farser/ast/node/terminal/ContainsNode.java
@@ -1,7 +1,7 @@
 package com.mmm.his.cer.utility.farser.ast.node.terminal;
 
+import com.mmm.his.cer.utility.farser.ast.node.LtrExpressionIterator;
 import com.mmm.his.cer.utility.farser.ast.node.type.BooleanExpression;
-import com.mmm.his.cer.utility.farser.ast.node.type.LtrExpressionIterator;
 import java.util.Collection;
 
 /**
