+++ src/main/java/com/mmm/his/cer/utility/farser/ast/node/type/NonTerminal.java
@@ -1,5 +1,7 @@
 package com.mmm.his.cer.utility.farser.ast.node.type;
 
+import com.mmm.his.cer.utility.farser.ast.node.LtrExpressionIterator;
+
 /**
  * This class represents a non-terminal node in the AST. These types of nodes will have a left and a
  * right child.
